Loupeview now handles panning color transforms for display are created only once.

+++ sqlkeywordmodel.cpp
@@ -2,16 +2,18 @@
 
 #include "sqlkeywordmodel.h"
 
-SqlKeywordModel::SqlKeywordModel(QObject* parent) : QAbstractItemModel(parent)
+SqlKeywordModel::SqlKeywordModel(QObject* parent) :
+    QAbstractItemModel(parent),
+    mRootItem(NULL)
 {
-
     // Construct the file tree
     createKeywordItems();
 }
 
 SqlKeywordModel::~SqlKeywordModel()
 {
-    deleteKeywordItems(mRootItem);
+    if (mRootItem != NULL)
+        deleteKeywordItems(mRootItem);
 }
 
 void SqlKeywordModel::createKeywordItems()
@@ -24,29 +26,30 @@ void SqlKeywordModel::createKeywordItems()
     // first create the root item
     QSqlQuery query;
 
-    QString queryText = QString("select id,keyword,parent_id from keyword where parent_id is NULL");
+    QString   queryText = QString("select id,keyword,parent_id from keyword where parent_id is NULL");
 
     if (!query.exec(queryText))
         qDebug() << query.lastError();
 
-    if (!query.first())
-        qDebug() << "No records";
-
-    mRootItem = new KeywordItem(query.value(0).toLongLong(),query.value(1).toString(),query.value(2).toLongLong());
-
-    createKeywordItemsRec(mRootItem);
+    if (query.first())
+    {
+        mRootItem = new KeywordItem(query.value(0).toLongLong(),query.value(1).toString(),query.value(2).toLongLong());
+        createKeywordItemsRec(mRootItem);
+    }
 }
 
 bool KComp(const KeywordItem* const & a, const KeywordItem* const & b)
 {
     int c = QString::compare(a->keyword, b->keyword,Qt::CaseInsensitive);
+
     return c < 0;
 }
 
 void SqlKeywordModel::createKeywordItemsRec(KeywordItem* root)
 {
     QSqlQuery query;
-    QString queryText = QString("select id,keyword,parent_id from keyword where parent_id = :parent_id");
+    QString   queryText = QString("select id,keyword,parent_id from keyword where parent_id = :parent_id");
+
     query.prepare(queryText);
 
     KeywordItem* item;
@@ -56,7 +59,7 @@ void SqlKeywordModel::createKeywordItemsRec(KeywordItem* root)
 
     while (query.next())
     {
-        item = new KeywordItem(query.value(0).toLongLong(),query.value(1).toString(), query.value(1).toLongLong());
+        item         = new KeywordItem(query.value(0).toLongLong(),query.value(1).toString(), query.value(1).toLongLong());
         item->parent = root;
         root->children.append(item);
         createKeywordItemsRec(item);
@@ -68,6 +71,7 @@ void SqlKeywordModel::createKeywordItemsRec(KeywordItem* root)
 void SqlKeywordModel::deleteKeywordItems(KeywordItem* root)
 {
     KeywordItem* item;
+
     foreach(item, root->children)
     {
         deleteKeywordItems(item);
@@ -117,7 +121,10 @@ int SqlKeywordModel::rowCount(const QModelIndex &parent) const
     else
         item = static_cast<KeywordItem*>(parent.internalPointer());
 
-    return item->children.size();
+    if (item != NULL)
+        return item->children.size();
+    else
+        return 0;
 }
 
 int SqlKeywordModel::columnCount(const QModelIndex & /*parent*/) const
